# docs/openapi-spec.yaml
openapi: 3.0.0
info:
  title: FuelSync Hub API
  version: 1.0.0
servers:
- url: /api/v1
paths:
...
  /api/v1/alerts/{id}:
    delete:
      summary: Delete an alert by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Alert deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
  /api/v1/alerts/{id}/read:
    patch:
      summary: Mark alert read
      responses:
        '200':
          $ref: '#/components/responses/Success'
        default:
          $ref: '#/components/responses/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
...